Adding choke points to Cistern in Runewick
Since the change in wrestling's damage has been made, it brings to light a flaw in the map design on the cisterns in Runewick. The map allows for all spawned rats on the first level to swarm and kill the newbie character.  I have added extra rocks etc to provide some saftey and choke points.